+++ src/api/apps/edgepay/edgepay.controller.ts
@@ -1,17 +1,17 @@
 import { Controller, Post, Body, BadRequestException } from "@nestjs/common"
-import { DwalletService } from "./dwallet.service"
-import { CreateTransactionDto } from "./dto/dwallet-tx"
+import { EdgepayService } from "./edgepay.service"
+import { CreateTransactionDto } from "./dto/edgepay-tx"
 import { statusMessages } from "../../../constants/statusMessages"
 import { ApiKeyAuthorizer } from "src/authorization/apikeyauthorizer/apikeyauthorizer.decorator"
 
-@Controller("dwallet")
-export class DwalletController {
-  constructor(private readonly dwalletService: DwalletService) { }
+@Controller("edgepay")
+export class EdgepayController {
+  constructor(private readonly edgepayService: EdgepayService) { }
 
   @Post("createtx")
   async createTransaction(@ApiKeyAuthorizer() userId: string, @Body() createTransactionDto: CreateTransactionDto) {
     try {
-      await this.dwalletService.createTransaction(createTransactionDto, userId)
+      await this.edgepayService.createTransaction(createTransactionDto, userId)
       return true
     }