+++ src/config/subscriptionConfig.ts
@@ -11,7 +11,7 @@ export const subscriptionConfig = {
 
 export const apiPricing = {
   airlake: 2,
-  dwallet: 5,
+  edgepay: 5,
   frostlake: 2,
   snowlake: 50,
   swapstream: 10,